Split SoftAssetReferenceEditorUtils
