+++ Modules/Resource/Editor/SoftAssetReferenceEditorUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfebb5a3fbe9f624d925bce9c45c40a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
